requirements.txt
gpt4all==1.0.8
python-dotenv==1.0.0
requests==2.31.0
tmdbv3api==1.9.0
aiohttp==3.9.1
colorlog==6.8.0 